+++ demo/dryvr_demo/thermo_demo2.py
@@ -1,6 +1,7 @@
 from origin_agent import thermo_agent
 from verse import Scenario
 from verse.plotter.plotter2D import *
+from verse.scenario.scenario import ScenarioConfig
 from verse.sensor.example_sensor.thermo_sensor import ThermoSensor
 import plotly.graph_objects as go
 from enum import Enum, auto
@@ -13,7 +14,8 @@ class ThermoMode(Enum):
 
 if __name__ == "__main__":
     input_code_name = './demo/dryvr_demo/thermo_controller.py'
-    scenario = Scenario()
+    config = ScenarioConfig(parallel=False)
+    scenario = Scenario(config)
 
     car = thermo_agent('test', file_name=input_code_name)
     scenario.add_agent(car)
