+++ verse/sensor/base_sensor.py
@@ -51,7 +51,7 @@ def add_states_3d(cont, disc, thing, val, cont_var, disc_var, stat_var):
 
 class BaseSensor():
     # The baseline sensor is omniscient. Each agent can get the state of all other agents
-    def sense(self, scenario, agent: BaseAgent, state_dict, lane_map):
+    def sense(self, agent: BaseAgent, state_dict, lane_map):
         cont = {}
         disc = {}
         len_dict = {'others': len(state_dict)-1}
