+++ verse/sensor/example_sensor/single_sensor.py
@@ -48,7 +48,7 @@ def add_states_3d(cont, disc, thing, val, cont_var, disc_var, stat_var):
 
 class SingleSensor():
     # The baseline sensor is omniscient. Each agent can get the state of all other agents
-    def sense(self, scenario, agent, state_dict, lane_map):
+    def sense(self, agent, state_dict, lane_map):
         cont = {}
         disc = {}
         len_dict = {'others': len(state_dict)-1}
@@ -125,7 +125,7 @@ class SingleSensor():
         return cont, disc, len_dict
 
 class FakeSensor3:
-    def sense(self, scenario, agent, state_dict, lane_map):
+    def sense(self, agent, state_dict, lane_map):
         cont = {}
         disc = {}
         len_dict = {'others': len(state_dict)-1}
@@ -174,7 +174,7 @@ def add_states_3d_ball(cont, disc, thing, val):
 
 
 class FakeSensor4:
-    def sense(self, scenario, agent, state_dict, lane_map):
+    def sense(self, agent, state_dict, lane_map):
         cont = {}
         disc = {}
         len_dict = {'others': len(state_dict)-1}
