+++ verse/sensor/example_sensor/thermo_sensor.py
@@ -42,7 +42,7 @@ def add_states_3d_ball(cont, disc, thing, val):
 
 
 class ThermoSensor():
-    def sense(self, scenario, agent, state_dict, lane_map):
+    def sense(self, agent, state_dict, lane_map):
         cont = {}
         disc = {}
         len_dict = {'others': len(state_dict)-1}
